nonebot_plugin_clash/__init__.py
from nonebot.plugin import PluginMetadata, inherit_supported_adapters, require

require("nonebot_plugin_alconna")
require("nonebot_plugin_htmlrender")

from . import __main__ as __main__  # noqa: E402
from .config import ConfigModel  # noqa: E402

__version__ = "0.1.0"
__plugin_meta__ = PluginMetadata(
    name="Clash",
    description="在 NoneBot 中控制你的 Clash",
    usage="待补充",
    type="application",
    homepage="https://github.com/lgc-NB2Dev/nonebot-plugin-clash",
    config=ConfigModel,
    supported_adapters=inherit_supported_adapters("nonebot_plugin_alconna"),
    extra={"License": "MIT", "Author": "student_2333"},
)
