pyproject.toml
[project]
name = "nonebot-plugin-clash"
version = "0.1.0"
description = "Control your clash core using NoneBot2"
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "pydantic>=1.10.4,<2",
    "nonebot-plugin-alconna>=0.33.8",
    "yarl>=1.9.4",
    "nonebot-plugin-htmlrender>=0.2.2",
    "jinja2>=3.1.2",
    "websockets>=12.0",
    "httpx>=0.25.2",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-clash"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_clash/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
